@@ -55,31 +55,27 @@
     "victory": "^32.0.2"
   },
   "resolutions": {
-    "hoist-non-react-statics": "3.0.1",
     "rebass": "4.0.7",
     "styled-components": "5.1.1",
     "styled-system": "5.1.5",
-    "js-yaml": "^3.13.1",
-    "https-proxy-agent": "^2.2.3",
-    "minimist": "^1.2.6",
-    "node-fetch": "^2.6.7",
     "lodash": "^4.17.21"
   },
   "devDependencies": {
-    "@svgr/webpack": "^6.5.0",
-    "@testing-library/cypress": "^9.0.0",
+    "@svgr/webpack": "^8.1.0",
+    "@testing-library/cypress": "^10.0.0",
     "@welldone-software/why-did-you-render": "^7.0.1",
-    "cypress": "^12.6.0",
-    "cypress-recurse": "^1.27.0",
-    "eslint": "^8.22.0",
-    "eslint-config-next": "^12.2.5",
-    "eslint-plugin-cypress": "^2.12.1",
-    "glob": "^8.0.3",
+    "cypress": "^13.2.0",
+    "cypress-recurse": "^1.35.2",
+    "eslint": "^8.49.0",
+    "eslint-config-next": "^13.4.19",
+    "eslint-plugin-cypress": "^2.14.0",
+    "glob": "^10.3.4",
     "imap-simple": "^5.1.0",
-    "jsdom": "^21.1.0",
-    "mailparser": "^3.6.3",
+    "jsdom": "^22.1.0",
+    "mailparser": "^3.6.5",
+    "msw": "^1.3.0",
     "mustache": "^4.2.0",
-    "nodemailer": "^6.9.1",
+    "nodemailer": "^6.9.5",
     "start-server-and-test": "^2.0.0"
   },
   "scripts": {
@@ -91,8 +87,9 @@
     "lint": "next lint",
     "start:testServer": "NODE_ENV=test yarn start",
     "build:test": "NODE_ENV=test yarn build",
-    "test:e2e": "yarn build:test && start-server-and-test start:testServer http://localhost:3100 cypress:run",
+    "cy:run": "NODE_ENV=test yarn run dev & cypress open",
     "cypress:run": "cypress run",
+    "test:e2e": "yarn build:test && start-server-and-test start:testServer http://localhost:3100 cypress:run",
     "test": "yarn run test:e2e",
     "script:build-translations": "node ./scripts/build-translations.js",
     "script:extract-i18n-keys": "yarn build && node ./scripts/extract-react-intl-keys.js",
@@ -100,5 +97,8 @@
     "git:getCommitSHA": "git rev-parse HEAD",
     "git:getCommitSHA:short": "git rev-parse --short HEAD",
     "git:getCommitRef": "git symbolic-ref HEAD"
+  },
+  "msw": {
+    "workerDirectory": "public"
   }
 }
